collector/requirements.txt
wget
pandas
flask
flask-cors
psycopg2-binary
pytest